Break a cycle
Once the connection is closed, remove the references from the protocol to the
reader and writer objects. It will effectively break a references cycle and
allow a large group of objects to be freed thanks to the reference counting
mechanism, instead of be garbage collected.

This obviously a really trivial patch, but in a large application emitting
a large amount of requests, it reduces the working time of the gc, and thus,
the latency.