Don't send full URL in tunnelled HTTPS requests
Ordinarily, the ProxyConnector rewrites req.path to include the full URL
of the request. However, when tunnelling HTTPS through an HTTP proxy, this
is unnecessary (because the CONNECT command tells the proxy which server
to connect to) and can cause problems (because some servers deal badly
with full URLs in GET requests). Therefore, we *shouldn't* rewrite the
path when making an HTTPS connection.