Improve performance of transport in-use list.
Connector keeps its in-use transport in a dictionary `_acquired`
mapping hostnames/ports/ssl to lists of transports.

Applications that open many connections to the same remote web service
end up having a single-item dictionary with a long list as value.

Whenever a transport is released, the item is `.remove()`d from the list
that has a O(n) complexity. Also `.append()` on a python list can have
O(n) complexity.

In this patch we use a set instead, so `.remove()` and `add()` have O(1)