[Backport][3.9] Only install async-timeout for Python < 3.11 (#7556) (#7558)
Replace `async-timeout` with `asyncio.timeout` for Python 3.11+.

Overall the change isn't as clean as I would have liked it to be. That's
mostly a result of the name conflict with `timeout`. I'm open for
suggestions on how to improve it.

`async-timeout` won't be a dependency for Python 3.11+ any longer.

Fixes: #7502
(cherry picked from commit db2c274c91d17eb410225d504c83d4bd6971d0ae)