Support brotlicffi alternatively to brotli (#7611)
## What do these changes do?

Support the brotlicffi implementation as an alternative to brotli. This
is necessary since the current version of brotli (as of 1.1.0) does not
work on PyPy. This uses the approach of preferring brotlicffi over
brotli in imports, and listing brotli or brotlicffi depending on the
Python implementation in dependencies, as recommended in brotlicffi
documentation:

https://pypi.org/project/brotlicffi/1.1.0.0/#using-brotlicffi-in-projects